refactor(filter): Rename substitutions mod to redactions
