+++ crates/snapbox/src/filter/redactions.rs
